// web3-reputation-client/src/model/web3ReputationAPI/nftListItemResponse.ts
/**
 * Generated by orval v6.12.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ServiceProviderResponse } from "./serviceProviderResponse";

export interface NftListItemResponse {
  tokenId?: number;
  rankLevel?: number;
  serviceProvider?: ServiceProviderResponse;
  xp?: number;
  rankName?: string;
  rankClaimable?: boolean;
  claimable?: boolean;
}